+++ src/API/useDownloadLinks.js
@@ -7,7 +7,8 @@ import { useOkapiKy } from '@folio/stripes/core';
 export const useDownloadLinks = (id) => {
   const ky = useOkapiKy();
 
-  const { data } = useQuery('error-download', {
+  const { data } = useQuery({
+    queryKey: ['error-download', id],
     queryFn: () => ky.get(`data-export-spring/jobs/${id}`).json(),
     enabled: !!id,
   });