+++ src/components/BulkEditList/BulkEditListResult/Preview/Preview.js
@@ -3,21 +3,31 @@ import { useParams, useLocation } from 'react-router';
 import { useIntl, FormattedMessage } from 'react-intl';
 import {
   Headline,
-  Accordion,
   AccordionSet,
-  Col,
   AccordionStatus,
-  Row,
   MessageBanner,
 } from '@folio/stripes/components';
 import { PreviewAccordion } from './PreviewAccordion';
-import { useDownloadLinks } from '../../../../API';
+import { ErrorsAccordion } from './ErrorsAccordion';
+import { useDownloadLinks, useErrorsList, usePreviewRecords } from '../../../../API';
 
 export const Preview = () => {
   const intl = useIntl();
   const location = useLocation();
   const { id } = useParams();
   const { data } = useDownloadLinks(id);
+  const { errors } = useErrorsList(id);
+  const { users } = usePreviewRecords(id);
+
+  const mappedErrors = errors?.map(e => {
+    const [identifier, message] = e.message.split(',');
+
+    return {
+      ...e,
+      identifier,
+      message,
+    };
+  });
 
   const processed = data?.progress?.processed;
   const fileUploadedName = useMemo(() => new URLSearchParams(location.search).get('fileName'), [location.search]);
@@ -25,43 +35,37 @@ export const Preview = () => {
   const title = useMemo(() => {
     const queryText = new URLSearchParams(location.search).get('queryText');
 
-    return fileUploadedName
-      ? intl.formatMessage({ id: 'ui-bulk-edit.preview.file.title' }, { fileUploadedName })
-      : intl.formatMessage({ id: 'ui-bulk-edit.preview.query.title' }, { queryText });
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [location.search]);
+    if (queryText) return intl.formatMessage({ id: 'ui-bulk-edit.preview.query.title' }, { queryText });
+
+    if (fileUploadedName) return intl.formatMessage({ id: 'ui-bulk-edit.preview.file.title' }, { fileUploadedName });
+
+    return null;
+  }, [fileUploadedName]);
 
   return (
     <AccordionStatus>
       {processed && (
+      <Headline size="large" margin="small">
         <MessageBanner type="success" contentClassName="SuccessBanner">
           <FormattedMessage
             id="ui-bulk-edit.recordsSuccessfullyChanged"
             values={{ value: processed }}
           />
-        </MessageBanner>)
-      }
-      <Headline size="large" margin="medium" tag="h3">
-        {title}
+        </MessageBanner>
       </Headline>
+      )}
+      {title && (
+        <Headline size="large" margin="medium">
+          {title}
+        </Headline>
+      )}
       <AccordionSet>
-        <PreviewAccordion />
-        {/* <Accordion label={intl.formatMessage({ id: 'ui-bulk-edit.list.errors.title' })}>
-          <Row>
-            <Col xs={12}>
-              <Headline size="medium" margin="xx-small" tag="h4">
-                  User-UUIDs.csv: 302 entries * 300 records matched
-              </Headline>
-            </Col>
-          </Row>
-
-          <Row>
-            <Col xs={12}>
-             Errors
-            </Col>
-          </Row>
-
-        </Accordion> */}
+        <PreviewAccordion users={users} />
+        <ErrorsAccordion
+          errors={mappedErrors}
+          entries={data?.progress?.total}
+          matched={data?.progress?.processed}
+        />
       </AccordionSet>
     </AccordionStatus>
   );